requirements.txt
Flask==2.2.5
pycryptodome==3.19.0
python-magic-bin==0.4.14